Переписал отрисовку комментов в попапе с помощью map. Удалил модуль film-comments.js
