+++ src/view/film-details-view.js
@@ -1,3 +1,5 @@
+import dayjs from 'dayjs';
+
 export const createFilmDetailsTemplate = (film) => {
   const {
     comments,
@@ -19,7 +21,39 @@ export const createFilmDetailsTemplate = (film) => {
     },
   } = film;
 
-  console.log(film);
+  const createFilmCommentTemplate = (comment) => {
+    const commentDate = comment.date;
+    const diff = dayjs().diff(commentDate, 'day');
+    let commentDay = '';
+
+    switch (diff) {
+      case 0:
+        commentDay = 'today';
+        break;
+      case 1:
+        commentDay = 'yesterday';
+        break;
+      case 2:
+        commentDay = '2 days ago';
+        break;
+      default:
+        commentDay = commentDate;
+    }
+
+    return `<li class="film-details__comment">
+      <span class="film-details__comment-emoji">
+        <img src="./images/emoji/${comment.emotion}.png" width="55" height="55" alt="emoji-${comment.emotion}">
+      </span>
+      <div>
+        <p class="film-details__comment-text">${comment.message}</p>
+        <p class="film-details__comment-info">
+          <span class="film-details__comment-author">${comment.author}</span>
+          <span class="film-details__comment-day">${commentDay}</span>
+          <button class="film-details__comment-delete">Delete</button>
+        </p>
+      </div>
+    </li>`;
+  };
 
   return `<section class="film-details">
     <form class="film-details__inner" action="" method="get">
@@ -96,6 +130,7 @@ export const createFilmDetailsTemplate = (film) => {
         <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
 
         <ul class="film-details__comments-list">
+          ${comments.map((comment) => createFilmCommentTemplate(comment)).join('')}
         </ul>
 
         <div class="film-details__new-comment">
